requirements.txt
langchain==0.2.0
langchain-google-genai==3.2.0
google-generativeai==0.5.0
python-dotenv==1.0.0
